+++ src/fast_init.py
