--- src/pipelines/transformers.py
@@ -1,5 +1,0 @@
-from src.pipelines.pipeline import Pipeline
-
-
-class HF_Pipeline(Pipeline):
-    pass
